Move input property visiting back to UnitOfWork
