+++ subprojects/core/src/main/java/org/gradle/internal/service/scopes/DefaultWorkInputListeners.java
@@ -19,9 +19,9 @@ package org.gradle.internal.service.scopes;
 import org.gradle.internal.event.AnonymousListenerBroadcast;
 import org.gradle.internal.event.ListenerManager;
 import org.gradle.internal.execution.UnitOfWork;
+import org.gradle.internal.execution.UnitOfWork.InputPropertyType;
 import org.gradle.internal.execution.WorkInputListener;
 import org.gradle.internal.execution.WorkInputListeners;
-import org.gradle.internal.execution.fingerprint.InputFingerprinter;
 
 import java.util.EnumSet;
 
@@ -43,7 +43,7 @@ public class DefaultWorkInputListeners implements WorkInputListeners {
     }
 
     @Override
-    public void broadcastFileSystemInputsOf(UnitOfWork work, EnumSet<InputFingerprinter.InputPropertyType> relevantTypes) {
+    public void broadcastFileSystemInputsOf(UnitOfWork work, EnumSet<InputPropertyType> relevantTypes) {
         broadcaster.getSource().onExecute(work, relevantTypes);
     }
 }
