+++ subprojects/execution/src/main/java/org/gradle/internal/execution/WorkInputListener.java
@@ -16,7 +16,7 @@
 
 package org.gradle.internal.execution;
 
-import org.gradle.internal.execution.fingerprint.InputFingerprinter;
+import org.gradle.internal.execution.UnitOfWork.InputPropertyType;
 import org.gradle.internal.service.scopes.EventScope;
 import org.gradle.internal.service.scopes.Scope.Global;
 
@@ -34,5 +34,5 @@ public interface WorkInputListener {
      * @param work the identity of the unit of work to be executed
      * @param relevantTypes the file system inputs relevant to the task execution
      */
-    void onExecute(UnitOfWork work, EnumSet<InputFingerprinter.InputPropertyType> relevantTypes);
+    void onExecute(UnitOfWork work, EnumSet<InputPropertyType> relevantTypes);
 }
