+++ subprojects/execution/src/main/java/org/gradle/internal/execution/WorkInputListeners.java
@@ -16,7 +16,6 @@
 
 package org.gradle.internal.execution;
 
-import org.gradle.internal.execution.fingerprint.InputFingerprinter;
 import org.gradle.internal.service.scopes.Scope.Global;
 import org.gradle.internal.service.scopes.ServiceScope;
 
@@ -34,5 +33,5 @@ public interface WorkInputListeners {
 
     void removeListener(WorkInputListener listener);
 
-    void broadcastFileSystemInputsOf(UnitOfWork work, EnumSet<InputFingerprinter.InputPropertyType> relevantTypes);
+    void broadcastFileSystemInputsOf(UnitOfWork work, EnumSet<UnitOfWork.InputPropertyType> relevantTypes);
 }
