+++ subprojects/execution/src/main/java/org/gradle/internal/execution/fingerprint/InputFingerprinter.java
@@ -18,17 +18,12 @@ package org.gradle.internal.execution.fingerprint;
 
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedMap;
-import org.gradle.api.file.FileCollection;
-import org.gradle.api.tasks.FileNormalizer;
+import org.gradle.internal.execution.UnitOfWork.InputVisitor;
 import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint;
-import org.gradle.internal.fingerprint.DirectorySensitivity;
 import org.gradle.internal.fingerprint.FileCollectionFingerprint;
-import org.gradle.internal.fingerprint.LineEndingSensitivity;
 import org.gradle.internal.snapshot.ValueSnapshot;
 
-import javax.annotation.Nullable;
 import java.util.function.Consumer;
-import java.util.function.Supplier;
 
 public interface InputFingerprinter {
     Result fingerprintInputProperties(
@@ -38,103 +33,6 @@ public interface InputFingerprinter {
         ImmutableSortedMap<String, CurrentFileCollectionFingerprint> knownCurrentFingerprints,
         Consumer<InputVisitor> inputs
     ) throws InputFingerprintingException, InputFileFingerprintingException;
-
-    interface InputVisitor {
-        default void visitInputProperty(
-            String propertyName,
-            ValueSupplier value
-        ) {}
-
-        default void visitInputFileProperty(
-            String propertyName,
-            InputPropertyType type,
-            FileValueSupplier value
-        ) {}
-
-    }
-
-    enum InputPropertyType {
-        /**
-         * Non-incremental inputs.
-         */
-        NON_INCREMENTAL(false, false),
-
-        /**
-         * Incremental inputs.
-         */
-        INCREMENTAL(true, false),
-
-        /**
-         * These are the primary inputs to the incremental work item;
-         * if they are empty the work item shouldn't be executed.
-         */
-        PRIMARY(true, true);
-
-        private final boolean incremental;
-        private final boolean skipWhenEmpty;
-
-        InputPropertyType(boolean incremental, boolean skipWhenEmpty) {
-            this.incremental = incremental;
-            this.skipWhenEmpty = skipWhenEmpty;
-        }
-
-        public boolean isIncremental() {
-            return incremental;
-        }
-
-        public boolean isSkipWhenEmpty() {
-            return skipWhenEmpty;
-        }
-    }
-
-    interface ValueSupplier {
-        @Nullable
-        Object getValue();
-    }
-
-    class FileValueSupplier implements ValueSupplier {
-        private final Object value;
-        private final Class<? extends FileNormalizer> normalizer;
-        private final DirectorySensitivity directorySensitivity;
-        private final LineEndingSensitivity lineEndingSensitivity;
-        private final Supplier<FileCollection> files;
-
-        public FileValueSupplier(
-            @Nullable Object value,
-            Class<? extends FileNormalizer> normalizer,
-            DirectorySensitivity directorySensitivity,
-            LineEndingSensitivity lineEndingSensitivity,
-            Supplier<FileCollection> files
-        ) {
-            this.value = value;
-            this.normalizer = normalizer;
-            this.directorySensitivity = directorySensitivity;
-            this.lineEndingSensitivity = lineEndingSensitivity;
-            this.files = files;
-        }
-
-        @Nullable
-        @Override
-        public Object getValue() {
-            return value;
-        }
-
-        public Class<? extends FileNormalizer> getNormalizer() {
-            return normalizer;
-        }
-
-        public DirectorySensitivity getDirectorySensitivity() {
-            return directorySensitivity;
-        }
-
-        public LineEndingSensitivity getLineEndingNormalization() {
-            return lineEndingSensitivity;
-        }
-
-        public FileCollection getFiles() {
-            return files.get();
-        }
-    }
 
     interface Result {
         /**
