+++ subprojects/launcher/src/test/groovy/org/gradle/tooling/internal/provider/continuous/ContinuousBuildActionExecutorTest.groovy
@@ -31,8 +31,7 @@ import org.gradle.internal.buildevents.BuildStartedTime
 import org.gradle.internal.buildtree.BuildActionRunner
 import org.gradle.internal.event.DefaultListenerManager
 import org.gradle.internal.execution.UnitOfWork
-import org.gradle.internal.execution.fingerprint.InputFingerprinter
-import org.gradle.internal.execution.fingerprint.InputFingerprinter.InputVisitor
+import org.gradle.internal.execution.UnitOfWork.InputVisitor
 import org.gradle.internal.invocation.BuildAction
 import org.gradle.internal.logging.text.TestStyledTextOutputFactory
 import org.gradle.internal.service.scopes.DefaultFileChangeListeners
@@ -53,6 +52,8 @@ import spock.util.concurrent.PollingConditions
 import java.util.concurrent.Executors
 import java.util.concurrent.Future
 import java.util.concurrent.TimeUnit
+
+import static org.gradle.internal.execution.UnitOfWork.InputPropertyType.PRIMARY
 
 @RedirectStdIn
 class ContinuousBuildActionExecutorTest extends ConcurrentSpec {
@@ -383,14 +384,14 @@ class ContinuousBuildActionExecutorTest extends ConcurrentSpec {
     }
 
     private void declareInput(File file) {
-        def valueSupplier = Stub(InputFingerprinter.FileValueSupplier) {
+        def valueSupplier = Stub(UnitOfWork.FileValueSupplier) {
             getFiles() >> TestFiles.fixed(file)
         }
         inputListeners.broadcastFileSystemInputsOf(Stub(UnitOfWork) {
             visitRegularInputs(_ as InputVisitor) >> { InputVisitor visitor ->
-                visitor.visitInputFileProperty("test", InputFingerprinter.InputPropertyType.PRIMARY, valueSupplier)
+                visitor.visitInputFileProperty("test", PRIMARY, valueSupplier)
             }
-        }, EnumSet.allOf(InputFingerprinter.InputPropertyType))
+        }, EnumSet.allOf(UnitOfWork.InputPropertyType))
     }
 
     private ContinuousBuildActionExecutor executer() {
